Use the location.speed value to show the current speed.
Oh no! I didn't the location.speed value at all. The calculation function (calculateApproximateSpeedKmH) was wasted.

+++ app/src/main/java/code/name/monkey/retromusic/activities/DriveModeActivity.kt
@@ -465,7 +465,7 @@ class DriveModeActivity : AbsMusicServiceActivity(), TextViewUpdateListener, Cal
         stopService(gpsRecordServiceIntent)
     }
 
-    override fun updateTextView(latitude: Double, longitude: Double, speed: Double) {
+    override fun updateTextView(latitude: Double, longitude: Double, speed: Float) {
         val formattedLatitude = String.format("%+013.8f", latitude)
         val formattedLongitude = String.format("%+013.8f", longitude)
         binding.gpsValue.setText("GPS Value\n$formattedLatitude\n$formattedLongitude")
